requirements.txt
django
djangorestframework
django-crontab
pandas
mysql-connector-python
openpyxl